requirements.txt
fastapi
uvicorn
graphene
PyJWT
prometheus-client
pydantic
pydantic_core 
pydantic-settings
pytest
requests
pytest-asyncio
pytest-cov
kafka
httpx